requirements.txt
openai>=1.3.5
python-dotenv==1.0.0
flask==3.0.0
flask-socketio==5.3.6
python-engineio==4.8.0
python-socketio==5.10.0
